app/requirements.txt
asgiref==3.5.2
click==8.1.3
dnspython==2.2.1
fastapi==0.68.2
h11==0.14.0
motor==3.1.1
pydantic==1.10.2
pymongo==4.3.3
python-decouple==3.6
pytz==2022.7
starlette==0.14.2
typing_extensions==4.4.0
uvicorn==0.15.0
